rename algorithm, fixed linking issue
