+++ src/software/ai/navigator/path_planner/hrvo/brute_force_nearest_neighbor_search.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <vector>
+
+template<class T, class F>
+static std::vector<T> nearestNeighbours(const T& this_robot, const std::vector<T>& input, double radius, F comparator);
