--- src/software/ai/navigator/path_planner/hrvo/fixed_radius_nearest_neighbor_search.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "extlibs/hrvo/hrvo_agent.h"
-
-std::vector<std::pair<double, double>> queryClosestNeighbors(size_t index, float radius, std::vector<std::pair<double, double>> &agents);
-
-template<class T, class F>
-static std::vector<T> nearestNeighbours(const T& this_robot, const std::vector<T>& input, double radius, F comparator);
